Rename training_type_plugin file to strategy
