+++ pytorch_lightning/strategies/__init__.py
@@ -10,5 +10,5 @@ from pytorch_lightning.strategies.sharded import DDPShardedStrategy  # noqa: F40
 from pytorch_lightning.strategies.sharded_spawn import DDPSpawnShardedStrategy  # noqa: F401
 from pytorch_lightning.strategies.single_device import SingleDeviceStrategy  # noqa: F401
 from pytorch_lightning.strategies.single_tpu import SingleTPUStrategy  # noqa: F401
+from pytorch_lightning.strategies.strategy import Strategy  # noqa: F401
 from pytorch_lightning.strategies.tpu_spawn import TPUSpawnStrategy  # noqa: F401
-from pytorch_lightning.strategies.training_type_plugin import Strategy  # noqa: F401
